requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
pinecone-client==5.0.1
neo4j==5.25.0
openai==1.51.0
groq==0.11.0
google-genai==0.3.0
tiktoken==0.8.0
pypdf==5.0.1
tenacity==9.0.0
loguru==0.7.2
